move internal kmp-tor-common classes/methods to package internal and prefix files with -
