+++ library/kmp-tor-common/src/commonMain/kotlin/io/matthewnelson/kmp/tor/common/address/OnionAddressV3.kt
@@ -19,7 +19,7 @@ import io.matthewnelson.component.encoding.base32.Base32
 import io.matthewnelson.component.encoding.base32.decodeBase32ToArray
 import io.matthewnelson.kmp.tor.common.annotation.ExperimentalTorApi
 import io.matthewnelson.kmp.tor.common.annotation.SealedValueClass
-import io.matthewnelson.kmp.tor.common.util.stripAddress
+import io.matthewnelson.kmp.tor.common.internal.stripAddress
 import kotlin.jvm.JvmInline
 import kotlin.jvm.JvmStatic
