+++ library/kmp-tor-common/src/commonMain/kotlin/io/matthewnelson/kmp/tor/common/clientauth/OnionClientAuthPublicKey_B32_X25519.kt
@@ -21,7 +21,7 @@ import io.matthewnelson.component.encoding.base32.Base32
 import io.matthewnelson.component.encoding.base32.decodeBase32ToArray
 import io.matthewnelson.kmp.tor.common.annotation.ExperimentalTorApi
 import io.matthewnelson.kmp.tor.common.annotation.SealedValueClass
-import io.matthewnelson.kmp.tor.common.util.descriptorString
+import io.matthewnelson.kmp.tor.common.internal.descriptorString
 import kotlin.jvm.JvmInline
 import kotlin.jvm.JvmStatic
